Move client class namespace to Carto::Http::Client #3778
