+++ lib/google_plus_api.rb
@@ -1,6 +1,6 @@
 require_relative './google_plus_api_user_data'
 require_relative './google_plus_config'
-require_relative 'carto/http_client'
+require_relative 'carto/http/client'
 
 class GooglePlusAPI
