+++ services/datasources/lib/datasources/url/gdrive.rb
@@ -1,7 +1,7 @@
 # encoding: utf-8
 
 require 'google/api_client'
-require_relative '../../../../../lib/carto/http_client'
+require_relative '../../../../../lib/carto/http/client'
 
 module CartoDB
   module Datasources
