+++ services/datasources/lib/datasources/url/mailchimp.rb
@@ -4,7 +4,7 @@ require 'json'
 require 'gibbon'
 require 'addressable/uri'
 require_relative '../base_oauth'
-require_relative '../../../../../lib/carto/http_client'
+require_relative '../../../../../lib/carto/http/client'
 
 module CartoDB
   module Datasources
