+++ services/named-maps-api-wrapper/lib/named-maps-wrapper/named_map.rb
@@ -1,7 +1,7 @@
 # encoding: utf-8
 
 require_relative '../../../../app/models/visualization/vizjson'
-require_relative '../../../../lib/carto/http_client'
+require_relative '../../../../lib/carto/http/client'
 
 module CartoDB
   module NamedMapsWrapper
