+++ services/sql-api/sql_api.rb
@@ -1,7 +1,7 @@
 # encoding: utf-8
 require 'open-uri'
 require 'json'
-require_relative '../../lib/carto/http_client'
+require_relative '../../lib/carto/http/client'
 
 module CartoDB
   class SQLApi
