+++ services/table-geocoder/lib/geocoder_cache.rb
@@ -1,5 +1,5 @@
 # encoding: utf-8
-require_relative '../../../lib/carto/http_client'
+require_relative '../../../lib/carto/http/client'
 
 module CartoDB
   class GeocoderCache
